reorganized carousel and escape-room-list components made some smaller components out of multi-used code can now properly edit and display escape rooms used in carousel and escape-room-list
