+++ client/src/assets/displayFromTemplate.css
@@ -1,6 +1,6 @@
 .create-new-display-box {
   width: 100%;
-  height: 90%;
+  height: 100%;
   margin: auto;
   display: flex;
   flex-flow: row;
@@ -37,21 +37,32 @@
 
   overflow-y: auto;
   display: flex;
-  flex-flow: row wrap;
-  justify-content: space-evenly;
+  flex-flow: column;
+  align-items: center;
 }
 .add-from-template.item {
-  max-width: 30%;
+  width: 90%;
+  min-height: 60%;
   text-align: center;
   border: 2px solid rgb(184, 184, 184);
   border-radius: 15px;
   overflow: hidden;
+
+  display: flex;
+  background-color: white;
+}
+.add-from-template.type-box {
+  height: 100%;
+  width: 15%;
+  max-width: 15%;
+  margin: 0 1%;
+  align-self: center;
 }
 .add-from-template.item-name {
-  margin: 2%;
-  min-height: 5%;
+  margin: 0;
   font-weight: bold;
-  font-size: 1.2em;
+  font-size: 1.5em;
+  transform: rotateZ(-90deg);
 }
 .add-from-template.footer {
   height: 8%;
@@ -85,9 +96,9 @@
   margin: 0;
 }
 .template-display.container {
-  height: 90%;
+  width: 100%;
   box-sizing: border-box;
-  background-color: rgb(253, 253, 253);
+  background-color: rgb(241, 241, 241);
   display: flex;
   flex-flow: column;
   justify-content: flex-start;
@@ -97,16 +108,16 @@
   text-align: center;
 }
 .template-display.content {
-  height: 90%;
+  height: 100%;
 }
 .template-display.misc {
   height: auto;
 }
 .template-display.title {
-  margin-bottom: 5%;
+  margin-bottom: 1%;
 }
 .template-display.subtitle {
-  margin-bottom: 5%;
+  margin-bottom: 2%;
 }
 
 .template-display.contents {
