+++ client/src/assets/edit.css
@@ -162,7 +162,7 @@
 }
 .close-edit {
   padding: 0 4% !important;
-  font-size: 1.4em;
+  font-size: 1.1em;
   display: flex;
   justify-content: center;
   align-items: center;
